removed long doubles

+++ src/moves/MultiParticle.h
@@ -45,7 +45,7 @@ private:
   std::vector<uint> moveType, moleculeIndex;
   const MoleculeLookup& molLookup;
 
-  long double GetCoeff();
+  double GetCoeff();
   void CalculateTrialDistRot();
   void RotateForceBiased(uint molIndex);
   void TranslateForceBiased(uint molIndex);
@@ -313,12 +313,12 @@ inline double MultiParticle::CalculateWRatio(XYZ const &lb_new, XYZ const &lb_ol
   return w_ratio;
 }
 
-inline long double MultiParticle::GetCoeff()
+inline double MultiParticle::GetCoeff()
 {
   // calculate (w_new->old/w_old->new) and return it.
   XYZ lbf_old, lbf_new; // lambda * BETA * force
   XYZ lbt_old, lbt_new; // lambda * BETA * torque
-  long double w_ratio = 1.0;
+  double w_ratio = 1.0;
   double lBeta = lambda * BETA;
   uint m, molNumber;
   double r_max = moveSetRef.GetRMAX(bPick);
@@ -358,12 +358,12 @@ inline void MultiParticle::Accept(const uint rejectState, const uint step)
 {
   // Here we compare the values of reference and trial and decide whether to
   // accept or reject the move
-  long double MPCoeff = GetCoeff();
+  double MPCoeff = GetCoeff();
   double uBoltz = exp(-BETA * (sysPotNew.Total() - sysPotRef.Total()));
-  long double accept = MPCoeff * uBoltz;
+  double accept = MPCoeff * uBoltz;
   double pr = prng();
   bool result = (rejectState == mv::fail_state::NO_FAIL) && pr < accept;
-  printf("%.15Lf, %.15lf, %.15Lf, %.15lf, %d\n", MPCoeff, uBoltz, accept, pr, result);
+  printf("%.15lf, %.15lf, %.15lf, %.15lf, %d\n", MPCoeff, uBoltz, accept, pr, result);
   if(result) {
     sysPotRef = sysPotNew;
     swap(coordCurrRef, newMolsPos);
